+++ src/osha/oira/ploneintranet/interfaces.py
@@ -1,0 +1,5 @@
+from zope.interface import Interface
+
+
+class IQuaiveForm(Interface):
+    """Marker interface for views that are designed to be embedded in Quaive"""